feat: allow specifying a custom media location fix: texcoords

+++ SezzUI/Modules/JobHud/Jobs/GNB.cs
@@ -20,7 +20,7 @@ namespace SezzUI.Modules.JobHud.Jobs
 			bar2.Add(new(bar2) {TextureActionId = 7531, CooldownActionId = 7531, StatusId = 1191, MaxStatusDuration = 20}); // Rampart
 			bar2.Add(new(bar2) {TextureActionId = 16140, CooldownActionId = 16140, StatusId = 1832, MaxStatusDuration = 20}); // Camouflage
 			bar2.Add(new(bar2) {TextureActionId = 16148, CooldownActionId = 16148, StatusId = 1834, MaxStatusDuration = 15}); // Nebula
-			bar2.Add(new(bar2) {TextureActionId = 16161, CooldownActionId = 16161, StatusIds = new[] {(uint) 1840, (uint) 2683}, MaxStatusDurations = new[] {7f, 8f}, GlowBorderStatusId = 1898}); // Heart of Stone
+			bar2.Add(new(bar2) {TextureActionId = 16161, CooldownActionId = 16161, StatusIds = new[] {1840u, 2683u}, MaxStatusDurations = new[] {7f, 8f}, GlowBorderStatusId = 1898}); // Heart of Stone
 			bar2.Add(new(bar2) {TextureActionId = 16152, CooldownActionId = 16152, StatusId = 1836, MaxStatusDuration = 10}); // Superbolide
 			hud.AddBar(bar2);
